Fixed param type checking & other bugfixes.
The parameter type checking routines for autogenerated INSERT/UPDATE/
DELETE queries was fixed. The tests were reconfigured to better catch
bad bind arrays. The arguments for DBALite_Statement::__construct()
were reorganized to promote use of the optional param type hint. The
bind checking routine was tweaked to autoconvert scalars to array.
This should work for converting objects to arrays too... as long as
the property names match the bind names. A few other bugs were fixed
as a result of the testing.