Revamp statement tests for multiple databases.
The DBALite_StatementTest class was moved into tests/DBALite/Statement/CommonTests.php
and files extending DBALite_Statement_CommonTests were added to tests/DBALite/Statement
for the different database drivers. This was each driver will get tested for statement
functionality. Tests for MySQL and PostrgeSQL were added. So far tests have been run
against MySQL but not PostgreSQL.

In order to accomodate this new testing scheme DBALite_Driver_SqliteTest now uses
the SqliteTest.sqlite DB instead of Select_Test_Db.sqlite. Two columns were added
to the SQLite scheme and two columns deleted from the MySQL schema. This resulted
in all of the DataSet files having to be updated.

The statement tests use the plumbing of the PHPUnit Database Extension to load and
truncate the test DB with data. The database operations are called directly rather
than using PHPUnit_Extension_Database_TestCase class.

In addition the test runners were unified into one file. The file will look to see
if a PDO driver is loaded and if so try to connect to the test DB (except SQLite)
and if that is successful it will add the Driver and Statment tests for that driver
to the test suite. The new file and class has been named DBALiteTestSuite.
Documentation on setting up the test databases and running the test suite has been
started in the docblock for DBALiteTestSuite.