Join condition tweaks, allow no direction in orderBy.
Some tweaks were made to DBALite_Select::join() to allow not passing
a join condition for joins that do not need it (NATURAL) and have it
not screw up the query formatting. If you need to pass columns to
join() use NULL for the condition.

DBALite_Select::orderBy() 2nd parameter $direction is now optional.
Not passing a value will not specify a direction and the database's
default will be used.

Added DBALite_DriverAbstract::getDriverName() in case you need to
determine which RDBMS brand you are connected to.

Updated tests for above changes and removed deprecated
PHPUnit_Framework_TestCase::assertType(). Testing now requires
PHPUnit 3.5.