Change to LGPL License, MySql UTF8 fix.
The project was changed to LGPL-3.0 or later license.
On PHP 5.3.6 and up the MySQL client/connection charset should be set in
the DSN string and not with SET NAMES. SET NAMES does not affect the
client side and PDO->quote could not work correctly b/c of mismatched
charsets. Also made small improvement in DBALite_Select::where().