Fix error in Mysql driver connection code and tweak unit tests.
Fixes Github #11.
Commit 2c55e1, which altered the way a Mysql connection was set to use UTF-8
to, taking advantage of improvements in the underlying C driver in PHP 5.3.6
and later. This commit contained a bug, which was not caught because I was
having trouble setting up DBUnit on my latest laptop.

Along with fixing this bug getting PHPUnit + DBUnit setup and working (I had
to kick PEAR to the curb and use Composer) forced me to tweak the connection
setting used for the test databases and I was able to fix a few more bugs
that I am not sure why they did not surface before.