Change way __call works on DriverAbstract and Statement classes.
Previously __call in DBALite_DriverAbstract and DBALite_Statement used a
whitelist of methods to pass down to the encapsulated PDO or PDOStatement.
A pull request was submitted that checked whether the method exists using
method_exists() (it actually used function_exists() which was wrong) so that
future methods added to PDO API can be exposed without any change to the
DBALite code. Closes github #12.